+++ CHANGELOG.md
@@ -9,10 +9,12 @@ Change
   ctx.success = (result, msg) => {
     data.message = msg || ''
     if(_.isPlainObject(result)){
-      data.data = result.data
-    }else{
-      data.data = result
+      if(_.has(result, data)){
+        data.data = result.data
+        return
+      }
     }    
+    data.data = result   
   }
   ```